requirements.txt
asgiref==3.7.2
Django==4.2.5
sqlparse==0.4.4
tzdata==2023.3

djangorestframework~=3.14.0
pillow
telebot~=0.0.5
celery~=5.3.4
telegram~=0.0.1